[Phase 7] 운동 기록 수정/삭제 기능 구현
- 백엔드: DELETE /api/workouts/:id 엔드포인트 추가 (CASCADE 삭제)
- 백엔드: PUT 엔드포인트에 duration_min 수정 지원 추가
- 백엔드: ExerciseLog.duration_min 타입 Numeric → SmallInteger 변경 및 마이그레이션
- API: updateWorkout 타입 정의, deleteWorkout 함수 추가
- Context: editWorkout, removeWorkout 메서드 추가
- 로그 페이지: 확장 카드 하단에 수정/삭제 버튼 + 삭제 확인 Dialog
- 수정 페이지: /workout/:id/edit 신규 구현 (LoggingStep editMode 재사용)
- 입력 정수화: 무게/횟수/시간(분) 필드 step=1, parseInt 적용, 표시값 Math.round 처리

+++ backend/app/models/workout.py
@@ -32,7 +32,7 @@ class ExerciseLog(Base):
     order_index: Mapped[int | None] = mapped_column(SmallInteger)
 
     # cardio / flexibility
-    duration_min: Mapped[Decimal | None] = mapped_column(Numeric(5, 1))
+    duration_min: Mapped[int | None] = mapped_column(SmallInteger)
     distance_km: Mapped[Decimal | None] = mapped_column(Numeric(6, 3))
     speed_kmh: Mapped[Decimal | None] = mapped_column(Numeric(4, 1))
     incline_pct: Mapped[Decimal | None] = mapped_column(Numeric(4, 1))
